Begin rewrite to use local scripts run remotely.
Previously rippling was done by copying the script and parameters
in their separate file to the destination host and then running
the script there.

Current practice is to have Backup/Backup as the working directory.
Tests written such that there is then no need to set PYTHONPATH.

After this commit, the plan is to work on a develop branch.